requirements.txt
inwx-domrobot==3.0.3
requests==2.22.0
python-dotenv 
